+++ workspaces/announcements/plugins/announcements-common/src/types.ts
@@ -62,10 +62,14 @@ export type Announcement = {
   active: boolean;
   /** Date indicating when the announcement starts (is visible to end users) */
   start_at: string;
+  /** Date indicating when the announcement should show until, using current filters. If omitted, the announcement is open-ended. */
+  until_date?: string | null;
   /** The team on whose behalf the announcement was published */
   on_behalf_of?: string;
   /** Array of tags associated with the announcement */
   tags?: Tag[];
+  /** Whether the notification is enabled */
+  sendNotification?: boolean;
 };
 
 /**
@@ -102,6 +106,10 @@ export type AnnouncementsFilters = {
   sortBy?: 'created_at' | 'start_at';
   /** Sorting order: "asc" for ascending or "desc" for descending */
   order?: 'asc' | 'desc';
+  /** Filter by current status (current date falls between start and until) */
+  current?: boolean;
+  /** Whether the notification is enabled */
+  sendNotification?: boolean;
 };
 
 /**