+++ workspaces/azure-storage-explorer/plugins/azure-storage/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-azure-storage-explorer
 
+## 0.6.0
+
+### Minor Changes
+
+- 4815dfd: Backstage version bump to v1.41.1
+
 ## 0.5.0
 
 ### Minor Changes