+++ workspaces/bookmarks/plugins/bookmarks/CHANGELOG.md
@@ -4,4 +4,22 @@
 
 ### Minor Changes
 
-- 80b21b2: Initial release
+- c344bc0: Allow links to be added to bookmarks that have different embed and "Open in new tab" URLs
+
+## 0.1.2
+
+### Patch Changes
+
+- 00058f7: Updated dependency `@mui/x-tree-view` to `8.11.1`.
+
+## 0.1.1
+
+### Patch Changes
+
+- 6597245: Updated dependency `@mui/x-tree-view` to `8.11.0`.
+
+## 0.1.0
+
+### Minor Changes
+
+- b01eb10: Initial release